update Bugs
Fix:
sorting problem. The fetchupdate was replacing it to what it used to be
(not sorted).
When you enter the information using add anime. It puts the number as a
string so have to check if the number is a string or number then cast
it. When checking if the new ep is up.